+++ src/dendromatics/ground.py
@@ -3,7 +3,7 @@ import warnings
 import CSF
 import numpy as np
 from scipy.interpolate import griddata
-from scipy.spatial import cKDTree
+from scipy.spatial import KDTree
 from sklearn.cluster import DBSCAN
 
 from .voxel.voxel import *
@@ -152,8 +152,8 @@ def clean_cloth(dtm_points):
             "accepted by clean_cloth().",
             stacklevel=2,
         )
-    tree = cKDTree(dtm_points[:, :2])
-    _, indexes = tree.query(dtm_points[:, :2], 15)
+    tree = KDTree(dtm_points[:, :2])
+    _, indexes = tree.query(dtm_points[:, :2], 15, workers=-1)
     abs_devs = np.abs(dtm_points[:, 2] - np.median(dtm_points[:, 2][indexes], axis=1))
     mads = np.median(abs_devs)
     clean_points = dtm_points[abs_devs <= 2 * mads]
@@ -225,8 +225,8 @@ def normalize_heights(cloud, dtm_points):
         Vector containing the normalized height values for the cloud points.
     """
 
-    tree = cKDTree(dtm_points[:, :2])
-    d, idx_pt_mesh = tree.query(cloud[:, :2], 3)
+    tree = KDTree(dtm_points[:, :2])
+    d, idx_pt_mesh = tree.query(cloud[:, :2], 3, workers=-1)
     # Z point cloud - Z dtm (Weighted average, based on distance)
     zs_diff_triples = cloud[:, 2] - np.average(
         dtm_points[:, 2][idx_pt_mesh], weights=d, axis=1